requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3
reportlab==4.0.7
numpy==1.26.1
python-dotenv==1.0.0
Werkzeug==3.0.1
